Add support for loading and saving json.
Remove the ability to change the element name dynamically (via ChildAsName or ValueAsName). This simplified reference parsing and therefore allowed for not writing out the dummy elements that cluttered up the whole data.

Reduce the number of children used for the description to a max of 10, reducing the crazy slowdown on large collections.

Add new context menu items for finer expand/collapse control.

Add the ability to set the values of a enumdef as a collection.
Add a message box to enum to display all the used values for that node throughout the entire document. This simplifies the creation of descriptions from existing data, or allows some metrics of the usage of each item.

Messages can now be selected and copy/pasted.