Improved loading of images.
Changed future to use a single processing thread, instead of a ton of worker threads that tie up the thread pool. The worker thread runs every 10ms so thats the minimum delay now.

Timelines now correctly sync the offset with the rest of the group.

Message is forced onto the main thread, fixing some crashes.

Added an 'add multiple' option to timelines with a single keyframe type thats only a file.